Redesigned the algorithm to clear format only for the selected part.
However, some html format can't be cleared, including headers (h1, h2,
...), tables, lists, indents, alignments and quotes. These can be
cleared or reverted by clicking the corresponding buttons in the toolbar
of the editor. For example, the headers can be reverted by the "Text"
item in the menu unfold by the "H" button in the toolbar. The "Clear
Format" button now mainly deal with fonts.